pyproject.toml
[project]
name = "realtime-phone-agent"
version = "0.1.0"
description = "Production-grade realtime phone agent using FastAPI, MLX, LangGraph, and Twilio"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.5.0",
    "python-multipart>=0.0.12",
    "httpx>=0.27.0",
    
    # ML/Audio
    "mlx>=0.18.0",
    "mlx-whisper>=0.4.0",
    "kokoro>=0.3.0",
    "soundfile>=0.12.1",
    "numpy>=1.26.0",
    "scipy>=1.14.0",
    
    # LLM/Agent
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    
    # Database
    "pinecone-client>=5.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    
    # Twilio
    "twilio>=9.3.0",
    
    # Monitoring
    "wandb>=0.18.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

